Tracks list in playlist

+++ app/src/main/java/ui/media/playlists/all/PlaylistViewHolder.kt
@@ -22,6 +22,6 @@ class PlaylistViewHolder(
         }
 
         tvPlaylistName.text = playlist.name
-        tvPlaylistAmountTracks.text = EndingConvertor.track(playlist.tracks.size.toLong())
+        tvPlaylistAmountTracks.text = EndingConvertor.track(playlist.tracks.size)
     }
 }
